Main_project/requirements.txt
# --- Core Framework ---
Flask==3.1.2
flask-cors==5.0.1
Werkzeug==3.1.3

# --- OCR & Image Processing ---
pytesseract==0.3.13
Pillow==10.3.0
opencv-python==4.10.0.84
numpy==1.26.4

# --- Machine Learning / Model Handling ---
scikit-learn==1.5.2
joblib==1.4.2

# --- Regex / Parsing / Utilities ---
regex==2024.9.11
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1

# --- Optional: QR / Barcode Support (for Aadhaar) ---
pyzbar==0.1.9
qrcode==7.4.2

# --- Optional: Logging / Debugging ---
tqdm==4.66.4
